fix bad_function_call error introduced by the missing of the deleter of Event unique_ptr

+++ main/lib/core/src/ExampleHardware.cc
@@ -84,7 +84,7 @@ namespace eudaq {
       // Zero suppressed data
       std::vector<unsigned char> data = ZeroSuppressEvent(MakeRawEvent(m_width, m_height), m_width);
       result.insert(result.end(), data.begin(), data.end());
-      unsigned short numhits = (result.size() - 8) / 6;
+      unsigned short numhits = (unsigned short)(result.size() - 8) / 6;
       setlittleendian<unsigned short>(&result[6], 0x8000 | numhits);
     }
     return result;
